plotter winput bug
I've found out that my program halts art the winput statement.

I've set a print statement before and after the line of code, and only the first line is printed in the console in jupyter!

    print('pre-winput')
    kast_width=winput( bind=input_raam , width=56, pos=scene.title_anchor, id=1)
    print('post-winput')

This line is located in the onderdelen/plotter.py  in function 'build scene' line 226.
I've checked if the function I'm calling with winput is called using a print statement (line 17) but it is not triggered.

I am using the vpython and extension versions as suggested by john (as can be seen in environments.yml).

Could this be a undiscovered bug in vpython ?